Reorganized files, now fetches from 50 pages
